Properly organise infeasible_problem_analysis lib
