+++ src/solver/infeasible-problem-analysis/CMakeLists.txt
@@ -1,16 +1,16 @@
 project(infeasible-problem-analysis)
 
 set(SRC_INFEASIBLE_PROBLEM_ANALYSIS
-        unfeasibility-analysis.h
-        constraint-slack-analysis.h
+        include/antares/solver/infeasible-problem-analysis/unfeasibility-analysis.h
+        include/antares/solver/infeasible-problem-analysis/constraint-slack-analysis.h
         constraint-slack-analysis.cpp
-        variables-bounds-consistency.h
+        include/antares/solver/infeasible-problem-analysis/variables-bounds-consistency.h
         variables-bounds-consistency.cpp
         unfeasible-pb-analyzer.cpp
-        unfeasible-pb-analyzer.h
-        report.h
+        include/antares/solver/infeasible-problem-analysis/unfeasible-pb-analyzer.h
+        include/antares/solver/infeasible-problem-analysis/report.h
         report.cpp
-        constraint.h
+        include/antares/solver/infeasible-problem-analysis/constraint.h
         constraint.cpp
 )
 
@@ -22,5 +22,9 @@ target_link_libraries(infeasible_problem_analysis
 )
 target_include_directories(infeasible_problem_analysis
         PUBLIC
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/..>
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+)
+
+install(DIRECTORY include/antares
+        DESTINATION "include"
 )
