+++ src/solver/infeasible-problem-analysis/report.cpp
@@ -18,8 +18,8 @@
  * You should have received a copy of the Mozilla Public Licence 2.0
  * along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
  */
-#include "report.h"
-#include "constraint.h"
+#include "antares/solver/infeasible-problem-analysis/report.h"
+#include "antares/solver/infeasible-problem-analysis/constraint.h"
 #include <antares/logs/logs.h>
 #include <algorithm>
 #pragma GCC diagnostic push
