+++ src/solver/infeasible-problem-analysis/variables-bounds-consistency.cpp
@@ -18,7 +18,7 @@
  * You should have received a copy of the Mozilla Public Licence 2.0
  * along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
  */
-#include "variables-bounds-consistency.h"
+#include "antares/solver/infeasible-problem-analysis/variables-bounds-consistency.h"
 #include <antares/logs/logs.h>
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wunused-parameter"
